fix: Fix edit site properties - EXO-66773 - Meeds-io/MIPs#72 (#256)
Prior to this change, when you open the drawer to update a new created site, the bannerUploadId is set to "0" and the API request param bannerRemoved is set to true resulting an exception for sites that have not set a new banner.
After this change, the bannerUploadId is set to null on each drawer opening.

+++ layout-management-services/src/main/java/org/exoplatform/layoutmanagement/rest/SiteManagementRestService.java
@@ -142,10 +142,10 @@ public class SiteManagementRestService implements ResourceContainer {
       portalConfig.setLabel(siteLabel);
       portalConfig.setDisplayed(displayed);
       portalConfig.setDisplayOrder(displayed ? displayOrder : 0);
-      if (bannerRemoved) {
-       layoutService.removeSiteBanner(siteName);
-       portalConfig.setBannerFileId(0);
-      } else {
+      if (bannerRemoved && portalConfig.getBannerFileId() != 0) {
+        layoutService.removeSiteBanner(siteName);
+        portalConfig.setBannerFileId(0);
+      } else if (StringUtils.isNotBlank(bannerUploadId)) {
         portalConfig.setBannerUploadId(bannerUploadId);
       }
       layoutService.save(portalConfig);
